+++ CHANGES.rst
@@ -11,6 +11,7 @@
   objects that have no value in the greenlet. See :issue:`1275`.
 - Fixed negative length in pywsgi's Input read functions for non chunked body.
   Reported in :issue:`1274` by tzickel.
+- Upgrabe libuv from 1.22.0 to 1.23.2.
 
 1.3.6 (2018-08-17)
 ==================